Add missing oslo.cache parameters
Since [1] was merged, ceilometer supports caching attributes of gnocchi
resource using the oslo.cache library. However the parameters to use
the feature are missing from ceilometer.conf auto-generated.

This change adds missing oslo.cache namespace so that parameters of
this library is properly included in the file generated by
the oslo-config-generator command.

[1] 2511cfb6e48c5d03cd198ecf9f09f36db3caced8

Change-Id: Ifa0c882f41e6a0cc56d7304e48e5479906fc1199