Fix gnocchi filter project for non-default domain
the keystoneclient calls used were wrong, and basically ignored
any filter_domain value, parsing all the projects.
This leads to failures to resolve project if when for example several
projects with the same name exist in different domains.

This patch adapts the resolution to proper calls - first resolve the
name of the domain to ID, than use that ID in the call to resolve the
project.

Change-Id: I91e509aa5427ee060e02591f719a815f71768618