Fix swift pollsters that use get_account
this is a followup to I38f9ff2bec0a2a3cb9dfc5c362284e33c12f3127

in swiftclient, head_account and get_account have somewhat different
signatures, specifically order of kwargs. While passing http_conn as
third arg works for head_account, it does not work for get_account.
In order for it to work for both, it must be passed as kwarg.

Change-Id: I5433dc248e3f02c67107466408ed8c9ea8dd16a7