Add user/project names to polled samples
Project and user names would be first fetched from cache, if not found,
they will be requested from keystone and then cached. Using cache will
significanlty reduce the number of calls made to keystone.

If ceilometer is configured with no caching backend, the results
will always be fetched by querying request to keystone.

A new config option, `tenant_name_discovery` is introduced
to operate this feature. This feature is optional and is disabled by default.

No attempts to identify names will be made if uuids are found to be `None`.

Signed-off-by: Yadnesh Kulkarni <ykulkarn@redhat.com>
Change-Id: Iee5dbf09a1fd3ac571746fc66d2683eb8e6a1b27