Add support to namespaces on dynamic pollsters
Problem description
===================
The hardcoded pollsters are defined by namespaces, so they
are instantied based on the namespaces provided to the
'AgentManager'.

The dynamic pollsters, on the other hand, are always instantied,
independent of the provided namespaces.

Proposal
========
To allow operators to define in which namespaces the dynamic
pollster will be deployed, we propose to add a new configuration
'namespaces' in the dynamic pollsters yaml configuration.
This configuration will support a single entry or a list of the
namespaces that the pollster must be instantiated.

Change-Id: I39ba0c3dd312a0601e02f8cfcab7a44e585a8a7f