Fix Non-OpenStack dynamic pollster credentials handling
There was a problem with the barbican credentials processing for
Non-OpenStack dynamic pollster. The credentials variables would be
retrieved as a byte string, which would break the processing.
Therefore, we needed to introduce a method to check if the credentials
are a String object. if they are not, we convert them to String.

Change-Id: I2084061eb8f543d4c557963599732e35cfa22996
(cherry picked from commit 37bf3efb287e755cff7ab4855ea03d6c15bf9752)