Properly handle 'resource_id' as None for Gnocchi publisher
This issue is related to [1]. In Python 3 (I did not check exactly in
which version), the following code was throwing an error
`data.sort(key=operator.attrgetter('resource_id'))`. Therefore, this
patch is proposing a solution for such situations to properly handle
samples where the `resource_id` is None. Of course, `resource_id`
cannot be None in Gnocchi. However, that can be properly handled by
removing the metric to be pushed from the `gnocchi_resource.yml` file.
Which is what we were doing when we proposed [1]. The problem is that
the sort is happening before that part of the code is executed, to
decide if the sample should be ignored or not.

[1] https://review.opendev.org/c/openstack/ceilometer/+/746717

Change-Id: Ie8eb42df3d5b9505160c9e9d6b86bdaa9a02d16a