Fix gnocchi install from git
Recent changes to git prevent git repos from being parsed if they
are not owned by the current user as part of a CVE fix. Some
project use stevedore to load plugins and this is broken if
gnocchi is installed from git via "sudo pip install".

This change uses devstacks git_clone to ensure that the gnocchi
repos is clonned under $DEST aka /opt/stack as the current user not
root. It then uses pip install -e to install the precloned repo.

To enable this GNOCCHI_REPO and GNOCCHI_BRANCH are intoduced
and the non standard GNOCCHI_GIT_PATH is removed.

A long standing comment to stop installing from git after
Gnocchi 4.0 is adressed by only installing from  git if
gnocchi is listed in LIBS_FROM_GIT. However since
4.4.2 is currently incompatible with openstacks upper-constraits
file the ci will continue to git however the git repo will
now be prepared by zuul using required-projects.

This will have the side effect of allowing
Depends on to work for PRs to gnocchi however testing that
is out of scope of this patch.

This patch adds a sample local.conf that can be used to deploy
locally for devopment.

Depends-On: https://review.opendev.org/c/openstack/telemetry-tempest-plugin/+/872350
Change-Id: I833ea3bffb18bec063423450be0f3b8ff92f9a5a