Add logs to help identify when a polling task finishes
During a troubleshooting session, we needed to discover the time
that it was taking for a polling task to finish. However, there
were no clear logs showing that. We could see when the task would
start, but not when it finished without errors. Therefore, we are
suggesting in this patch this small addition to the code base, which
can help operators to identify when a polling task has finished.

We added it in INFO log level because the message regarding the start
of the polling is also in INFO. However, we can change it to DEBUG if
the community thinks that DEBUG log level is better in this context.

Change-Id: I7358add1e45abe4a1288e51efa2f91d07c6acacd