Pin Gnocchi 4.5.1 to Zed release
Instead of setting Gnocchi version through localrc configuration, it
is hardcoded in devstack settings. A particular stable release
of OpenStack is not expected to be tested with multiple Gnocchi
releases and therefore the version pinned in this release is not expected to change.

Closes-Bug: #2035074
Change-Id: I60e9716555999f19d360b8219b9539c88622e07f
(cherry picked from commit f47af941f7134ddecf65b19f12d1a325ba54fded)