Add sg-core backend support in devstack plugin
This change also adds a new variable to set during
devstack deployment: "CEILOMETER_BACKENDS". Its
usage is similar to the "CEILOMETER_BACKEND". The difference
is, that CEILOMETER_BACKENDS expects a comma separated
list of backends (instead of only one backend). The
behavior of CEILOMETER_BACKEND is still the same for
backwards compatibility. If both CEILOMETER_BACKEND
and CEILOMETER_BACKENDS are set, then the value
of CEILOMETER_BACKEND is appended to CEILOMETER_BACKENDS.

The default backend if both "CEILOMETER_BACKEND" and
"CEILOMETER_BACKENDS" are empty is still gnocchi

Change-Id: Ib92fa062784479796c8a2ca4bd69a538e4e05f63