Fix devstack plugin for CEILOMETER_BACKEND=none
My recent changes to the devstack plugin broke the plugin for
situations, where CEILOMETER_BACKEND=none. The reason is the
addition of the CEILOMETER_BACKENDS variable, which uses
"," for delimiting individual backends. This meant, that
when using CEILOMETER_BACKEND=none, the CEILOMETER_BACKENDS
would in fact be ",none", which wouldn't equal to "none" in
a check in the plugin.sh

The fix is to delete all "," characters from the
CEILOMETER_BACKENDS before checking if it equals to "none"

Change-Id: Ief09bf1d5da79d7f9b7a2ba4055be2b2ef3dc1b8