Deprecate OpenContrail support
The current OpenContrail neutron plugin, which is available in
GitHub[1] is not compatible with the recent OpenStack releases, mainly
because of openstacksdk pinned to <2.

Because the plugin can't be used with recent releases of OpenStack, we
do not expect any users using it as their network backend in their
OpenStack deployment.

[1] https://github.com/tungstenfabric/tf-neutron-plugin

Change-Id: I66cec47a9238dab68445838e5599bfe1c45d2e4e