Fix dynamic pollsters on compute nodes and coordination process
Ceilometer can coordinate Pollsters based on the discovery process. The
idea (according to what we have in the code so far) is that the values
"discovered" by the discoveries are bound/linked to a discovery agent.
Then, only the pollsters on the node where these agents run are the ones
being able to process, as the code will only enable pollster to gather
data if the discovery returns some data to feed the pollsters.

Some pollsters, such as the Dynamic pollster system for non-OpenStack
API and for compute nodes do not depend on the discovery process. Thus,
they rely on a "local" node discovery process that fakes the discovery
and always returns a fixed value.

The problem in this situation is that multiple Dynamic pollsters in the
compute nodes are coordinated based on the discovery which is always the
same. Therefore, only one of them is going to execute. As a consequence
of that, data will be lost.

To address this situation we have an approach to make the local node
discovery return a unique "resource" based on the node where it runs.
Furthermore, we are introducing a process that will enable the
coordination of pollsters themselves.

Change-Id: I2a0a3bc17bc3a6d78e1629fd0a460d75589025dd