Add more metadata to volume.size and volume.backup.size pollster
When events are received from cinder, this includes the the volume_type_id
field (and not the volume_type name). Since Wallaby version it is now
possible to fetch this id from the API, so we can now make this poller
consistent with what is sent over the notification bus.

Another useful missing field for backups is the is_incremental field.

As the backup.user_id field was added in microversion 3.56, it makes sense
to add this information here as well to make it consistent with other
samples.

Closes-bug: #2064290
Change-Id: Ib109f399bab43ed8a6a0e0afea27ff0e5f7ae883