Enable instance metadata polling
When polling libvirt for instance data, we weren't getting
the instance metadata. In particular we weren't getting the
metering.server_group, which is useful for autoscaling.

The metadata need to be retreived by an API call to nova.
The code already called nova API to get flavor ID. I adjusted
the code to get the flavor ID and instance metadata by the same
call, so the number of API calls is unchanged.

Change-Id: I8034fca8a86263f38e396c4c1512546863a754db