Fix Prom exporter resource_name parsing
This changes the code to follow the original code in sg-core.
The resource_name label value in sg-core is being overwritten
when there is a "display name" and "name". It shouldn't be
duplicated.

It also removes the index variable, which is there in the
original implementation mostly to do this overwrite. I
don't think it's needed in this Python implementation.

Related sg-core code: https://github.com/infrawatch/sg-core/blob/master/plugins/handler/ceilometer-metrics/main.go#L269

Change-Id: I72971d51fa37b43e348605af9f6bb52cfa651369