Make all metrics be gauge type
The cumulative type metrics in Ceilometer return an already
incremented value, making the "inc()" function of the
Prometheus library to give fake values due to adding
already incremented values instead of the increment.

Moreover, the prometheus_client adds automatically
_total suffix to every counter metric, which kills the
need of this to be a drop-in replacement for sg-core
exporter metrics.

In this case, all metrics are of type Gauge as this
value is lost when the metric is stored in Prometheus,
which does not track the metric type.

Change-Id: Ib4f3df14849499a7fb6b40c98ac4a12040cdfd17