Remove stale metrics from prometheus exporter
If a VM is deleted from a node, it is necessary to remove its
samples from the exporter Prometheus data to avoid them to
become stale.

As unregistering a collector is a very cheap action (del from
a dict), we are unregistering every collector on each polling
cycle, thus guaranteeing that the only metrics exposed after
every polling cycle are the only valid ones and purging
the stale metrics.

Change-Id: I92dc47a9260630df6c4e94f6d79b4601d79a6b59