Do not drop floating IPs/VPNs with unknown status
Currently, Ceilometer checks the state of floating IPs and
Neutron VPNaaS VPNs and does not generate samples if they have
a status value of None, leaving a warning in the Ceilometer logs.
In addition, VPNs are also dropped if they have status ERROR.

There are two major issues with this:

* It makes visibility of resources with unknown/invalid/error
  state worse, since samples are not being generated for affected
  resources at all.
* Downstream rating/billing services such as CloudKitty prefer to know
  everything about as many resources as possible, so business logic can
  be implemented around that information. Sometimes resources should be
  billed even if they are in ERROR state, because they still consume
  resources.

This patch changes the ip.floating and network.services.vpn pollsters
to always sample all found floating IPs and VPNs. If they have status
ERROR, the sample volume is set to 7. For resources with status None
the sample volume is set to -1, same as when an unknown non-None state
is found.

Neutron FWaaS firewalls also implement filtering by state, but as those
meters are deprecated due to the retirement of Neutron FWaaS, they were
not fixed.

Change-Id: Ide197631f2fccdb88457159ddf37b023cb3b1f80