Get more flavor attributes from libvirt metadata (if available)
From 2025.2 Flamingo onwards, Nova now supplies the flavor ID
Ceilometer requires directly in the libvirt domain metadata for
an instance.

Implement the ability to fetch additional flavor attributes from
libvirt domain metadata, when available.

To maintain backwards compatibility, if the flavor ID is not found
in libvirt metadata, Ceilometer will continue to perform Nova API
queries to fetch it (with debug logs added to make it obvious when
it does so).

In addition to the required flavor ID, the extra specs for a flavor
is also now available in the libvirt domain metadata, so fetch these
and populate the ``extra_specs`` dictionary structure in compute
samples.

With the flavor ID now available in libvirt metadata, with the
exception of user-provided server metadata, Ceilometer no longer needs
to query Nova API for compute polling. To eliminate the need to query
Nova API entirely, a ``[compute]/fetch_extra_metadata`` option has
also been added to allow Ceilometer Compute Agent to be configured
to disable any metadata attributes that require Nova API queries.
If user metadata is still required, set this option to ``True``
(with the downside that per-instance queries will still be performed
to get this metadata).

``[compute]/fetch_extra_metadata`` will be set to ``True`` by default
initially, but after a couple of release cycles it should be changed to
``False`` by default to allow Ceilometer to be more performant and
lightweight out of the box.

Signed-off-by: Callum Dickinson <callum.dickinson@catalystcloud.nz>
Depends-On: https://review.opendev.org/c/openstack/nova/+/942974
Story: #2011371
Task: #51738
Change-Id: I57228ca951904c0f80ce6e644f2f9f51fb33fa04