Make the instance 'host' attribute optional
Make the ``host`` resource attribute on the ``instance`` resource type
optional in Gnocchi.

This allows the hypervisor a compute instance is running on to be
withheld from Gnocchi's resource metadata, which may be required
for security reasons e.g. for public clouds.

To handle updating existing attributes, changes have been made to
the ``upgrade_resource_types`` function to make it check if any
existing attributes are identical to the specified schema, and
update the resource type accordingly.

Change-Id: I26875b876be6401164442d2f5e52e2eb9b8e703c