Catch exceptions for interfaceStats in libvirt inspector
In the Libvirt inspector we call the interfaceStats()
function but there is a race between us checking if
the domain is stopped and calling interfaceStats.

This adds a try-catch around interfaceStats the
same way as is already done for blockStats etc.

Closes-Bug: #2113768
Change-Id: I25ba64f5eefe8140b45bc4ff7e20c323dc5efd79
(cherry picked from commit 70a56ba670157bf72d1513b8c177c9294071e2eb)
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>