Fix volume provider pool capacity metrics
Turns out there are a few small issues with the new volume metrics
when used with ceph as the storage backend. This change
fixes a few issues encountered with ceph. It also adds
a new storage pool in the tests, which triggered all of the issues.

List of issues this fixes:
- The max_over_subscription_ratio is returned as a string, which
caused an exception.
- pool_name is missing for some pools
- provisioned_capacity_gb is missing for some pools

Closes-Bug: #2113903
Change-Id: I4fd2fd89bd666c2466be5af2c15eb5134de0ce48