Add vcpus pollster
The vcpus meter would be a useful metric to use when getting
CPU usage for an instance from Gnocchi without needing to know
the number of vCPUs of the instance ahead of time, but the fact
that it is only supplied by notifications makes it difficult to
use for this use case.

Add a compute pollster for the vcpus meter using already available
metadata, to make sure the vcpus metric in Gnocchi gets published
at the same rate as other compute metrics.

Change-Id: If41d51088a2b6fc112f525fc30749d39fe341e46
Signed-off-by: Callum Dickinson <callum.dickinson@catalystcloud.nz>