Use 'actual' memory statistic for memory pollster
The memory pollster was configured to use the 'available' memory
statistic in libvirt, which represents the amount of usable memory
as seen inside the domain.

What it should be using is the 'actual' statistic, which is the amount
of memory actually allocated to the domain, to match the notification
meter. This patch fixes this discrepancy.

From what I can tell, the memory_usage InstanceStats value is correct
as-is and does not need changing.

Change-Id: I28d966fb65a81aa97890e68f3859c76d69601dd1
Signed-off-by: Callum Dickinson <callum.dickinson@catalystcloud.nz>