Add memory.available metric
Add a new memory.available compute pollster metric that exposes
the memory available as seen by the instance itself.

This is different from the memory metric, which tracks the amount
of raw memory allocated to an instance by the hypervisor (and matches
the amount of memory configured in the instance flavour).

The memory.available and memory.usage metrics can be used in
Gnocchi aggregate queries to get the RAM usage of an instance
as a percentage, without needing to know the amount of RAM
allocated to, or actually available in, the instance beforehand.

Change-Id: I3be6dea1e81238d5d70a0c48e35013d0d4ec4e9e
Signed-off-by: Callum Dickinson <callum.dickinson@catalystcloud.nz>