Fix removing stale metrics from Prometheus exporter
The Prometheus exporter tries to remove stale metrics every polling
cycle so VMs that have removed from the node do not leave metrics there.
This works fine when there are multiple VMs in the node, but if the
last VM is removed from the node, its metrics survive in the exporter
forever.

This is due Ceilometer not running the pollsters when there are no VMs
available to collect metrics and the current code running the cleanup on
the pollster code.

This fix moves the existent cleanup code so it can run even if the
pollster does not execute due to no VMs available, which removes stale
metrics in every case.

Change-Id: I8394c71a78f9b0004514fbb624ac7436d3c60e61
Signed-off-by: jlarriba <jlarriba@redhat.com>