Fix polling manager exceptions
The ceilometer.tests.unit.polling.test_manager.TestPollingAgent.test_start
test was throwing exceptions, which weren't caught by the tests because
of using threads, but they were visible in the test outputs. This patch
modifies a condition in the polling manager to prevent these exceptions
to happen in tests and in real environments as well.

Change-Id: Ib51fd788ac1259c891f4e23ce8430c96c5fe9179
Signed-off-by: Jaromir Wysoglad <jwysogla@redhat.com>